asr/model/requirements.txt
openai-whisper
setuptools-rust
fastapi
pydantic
